pyproject.toml
[project]
name = "02-scenariodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "YounghwanJoo", email = "neokein@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.115",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.30",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "scenario_db"

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=9.0.3",
    "testcontainers[postgresql]>=4.0",
]
notebook = [
    "jupyterlab>=4.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.20",
    "kaleido>=0.2",
    "networkx>=3.0",
    "tabulate>=0.9",
    "ipywidgets>=8.0",
    "python-dotenv>=1.2.2",
]
dashboard = [
    "streamlit>=1.40",
    "requests>=2.31",
]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
markers = [
    "unit: pure Python, no external deps",
    "integration: requires Docker + PostgreSQL testcontainer",
]
